Command autorun functionality added.
The original autoruns functionality would only run Modules, and there was no clear way to automatically run Commands.  Furthermore, Commands may not be called from Modules (and vise-versa) due to circular dependencies.

This new feature is a temporary workaround, however there is a desire to have modules capable of storing child files, which currently is not possible with the Store command.